refactor videometadata into its own file
